requirements.txt
selenium==3.141.0
numpy==1.19.4
pandas==1.1.4
beautifulsoup4==4.9.3
html5lib==1.1
lxml==4.6.2
python_dateutil==2.8.1
streamlit==0.74.1
